+++ commander/executor/create_create2transfer_commitments.go
@@ -72,7 +72,7 @@ func (t *TransactionExecutor) createC2TCommitment(
 
 		var transfers *AppliedC2Transfers
 
-		maxAppliedTransfers := t.cfg.TxsPerCommitment - uint64(len(appliedTransfers))
+		maxAppliedTransfers := t.cfg.TxsPerCommitment - uint32(len(appliedTransfers))
 		transfers, err = t.ApplyCreate2Transfers(pendingTransfers, maxAppliedTransfers)
 		if err != nil {
 			return nil, nil, err
@@ -90,7 +90,7 @@ func (t *TransactionExecutor) createC2TCommitment(
 			break
 		}
 
-		limit := pendingTxsCountMultiplier*t.cfg.TxsPerCommitment + uint64(len(appliedTransfers)+len(invalidTransfers))
+		limit := pendingTxsCountMultiplier*t.cfg.TxsPerCommitment + uint32(len(appliedTransfers)+len(invalidTransfers))
 		pendingTransfers, err = t.storage.GetPendingCreate2Transfers(limit)
 		if err != nil {
 			return nil, nil, err
