+++ commander/executor/create_create2transfer_commitments_test.go
@@ -210,7 +210,7 @@ func (s *Create2TransferCommitmentsTestSuite) addCreate2Transfers(transfers []mo
 	}
 }
 
-func (s *Create2TransferCommitmentsTestSuite) prepareAndReturnPendingCreate2Transfers(transfersAmount uint64) []models.Create2Transfer {
+func (s *Create2TransferCommitmentsTestSuite) prepareAndReturnPendingCreate2Transfers(transfersAmount uint32) []models.Create2Transfer {
 	transfers := generateValidCreate2Transfers(transfersAmount, &models.PublicKey{1, 2, 3})
 	s.addCreate2Transfers(transfers)
