+++ config/config.go
@@ -38,7 +38,7 @@ func GetConfig() *Config {
 		Rollup: &RollupConfig{
 			SyncSize:               getUint32("rollup.sync_size", 50),
 			FeeReceiverPubKeyID:    getUint32("rollup.fee_receiver_pub_key_id", 0),
-			TxsPerCommitment:       getUint64("rollup.txs_per_commitment", 32),
+			TxsPerCommitment:       getUint32("rollup.txs_per_commitment", 32),
 			MinCommitmentsPerBatch: getUint32("rollup.min_commitments_per_batch", 1),
 			MaxCommitmentsPerBatch: getUint32("rollup.max_commitments_per_batch", 32),
 			CommitmentLoopInterval: getDuration("rollup.commitment_loop_interval", 500*time.Millisecond),
@@ -76,7 +76,7 @@ func GetTestConfig() *Config {
 		Rollup: &RollupConfig{
 			SyncSize:               getUint32("rollup.sync_size", 50),
 			FeeReceiverPubKeyID:    getUint32("rollup.fee_receiver_pub_key_id", 0),
-			TxsPerCommitment:       getUint64("rollup.txs_per_commitment", 2),
+			TxsPerCommitment:       getUint32("rollup.txs_per_commitment", 2),
 			MinCommitmentsPerBatch: getUint32("rollup.min_commitments_per_batch", 1),
 			MaxCommitmentsPerBatch: getUint32("rollup.max_commitments_per_batch", 32),
 			CommitmentLoopInterval: getDuration("rollup.commitment_loop_interval", 500*time.Millisecond),
